+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Guide: https://keepachangelog.com/en/1.0.0/
 <!-- Add changes for active work here -->
 
 - [SearchResult] Mark `matchingName` field as deprecated and add note for absence of values in ApiType.searchBox results.
+- [SearchEngine] Add documentation and assertion that ApiType.searchBox does _not_ support batch requests.
 - [Demo] Add example for forward(query:options:completion:) to Demo app using SwiftUI.
 - [Search] Add forward(query:options:completion:) function to search forward/ API endpoint.