+++ CHANGELOG.md
@@ -8,8 +8,13 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 <!-- Add changes for active work here -->
 
+## 2.5.1
+
 - [Demo] Add example for forward(query:options:completion:) to Demo app using SwiftUI.
 - [Search] Add forward(query:options:completion:) function to search forward/ API endpoint.
+
+**MapboxCommon**: v24.7.0
+**MapboxCoreSearch**: v2.5.1
 
 ## 2.5.0